GMG: move preconditioner to header
