+++ include/aspect/simulator/solver/stokes_matrix_free.h
@@ -24,9 +24,12 @@
 #include <aspect/global.h>
 #include <aspect/parameters.h>
 #include <aspect/simulator/solver/interface.h>
+#include <aspect/utilities.h>
 
 #include <deal.II/multigrid/mg_transfer_matrix_free.h>
 #include <deal.II/multigrid/mg_transfer_global_coarsening.templates.h>
+
+#include <deal.II/lac/solver_bicgstab.h>
 
 /**
  * Typedef for the number type for the multigrid operators. Can be either float or double.
@@ -133,6 +136,250 @@ namespace aspect
   template <int dim>
   std::unique_ptr<StokesMatrixFreeHandler<dim>> create_matrix_free_solver(Simulator<dim> &simulator, const Parameters<dim> &parameters);
 
+
+  namespace internal
+  {
+
+    /**
+     * Implement the block Schur preconditioner for the Stokes system.
+     */
+    template <class StokesMatrixType, class ABlockMatrixType, class SchurComplementMatrixType,
+              class ABlockPreconditionerType, class SchurComplementPreconditionerType>
+    class BlockSchurGMGPreconditioner : public Subscriptor
+    {
+      public:
+        /**
+        * @brief Constructor
+        *
+        * @param Stokes_matrix The entire Stokes matrix
+        * @param A_block The A block of the Stokes matrix
+        * @param Schur_complement_block The matrix which describes the Schur complement approximation
+        * @param A_block_preconditioner Preconditioner object for the matrix A.
+        * @param Schur_complement_preconditioner Preconditioner object for the Schur complement.
+        * @param do_solve_A A flag indicating whether we should actually solve with
+        *     the matrix $A_block$, or only apply one preconditioner step with it.
+        * @param do_solve_Schur_complement A flag indicating whether we should actually solve with
+        *     the matrix $Schur_complement_block$, or only apply one preconditioner step with it.
+        * @param A_block_is_symmetric A flag indicating whether the A block is symmetric.
+        * @param A_block_tolerance The tolerance for the CG solver which computes
+        *     the inverse of the A block.
+        * @param Schur_complement_tolerance The tolerance for the CG solver which computes
+        *     the inverse of the Schur complement block (Schur complement approximation matrix).
+        */
+        BlockSchurGMGPreconditioner (const StokesMatrixType                  &Stokes_matrix,
+                                     const ABlockMatrixType                  &A_block,
+                                     const SchurComplementMatrixType         &Schur_complement_block,
+                                     const ABlockPreconditionerType          &A_block_preconditioner,
+                                     const SchurComplementPreconditionerType &Schur_complement_preconditioner,
+                                     const bool                               do_solve_A,
+                                     const bool                               do_solve_Schur_complement,
+                                     const bool                               A_block_is_symmetric,
+                                     const double                             A_block_tolerance,
+                                     const double                             Schur_complement_tolerance);
+
+        /**
+        * Matrix vector product with this preconditioner object.
+        */
+        void vmult (dealii::LinearAlgebra::distributed::BlockVector<double>       &dst,
+                    const dealii::LinearAlgebra::distributed::BlockVector<double> &src) const;
+
+        unsigned int n_iterations_A_block() const;
+        unsigned int n_iterations_Schur_complement() const;
+
+
+      private:
+        /**
+        * References to the various matrix object this preconditioner works on.
+        */
+        const StokesMatrixType                  &stokes_matrix;
+        const ABlockMatrixType                  &A_block;
+        const SchurComplementMatrixType         &Schur_complement_block;
+        const ABlockPreconditionerType          &A_block_preconditioner;
+        const SchurComplementPreconditionerType &Schur_complement_preconditioner;
+
+        /**
+        * Whether to actually invert the $\tilde M$ or $\tilde A$ of the preconditioner matrix
+        * or to just apply a single preconditioner step with it.
+        */
+        const bool                                                      do_solve_A;
+        const bool                                                      do_solve_Schur_complement;
+        const bool                                                      A_block_is_symmetric;
+        mutable unsigned int                                            n_iterations_A_;
+        mutable unsigned int                                            n_iterations_Schur_complement_;
+        const double                                                    A_block_tolerance;
+        const double                                                    Schur_complement_tolerance;
+        mutable dealii::LinearAlgebra::distributed::BlockVector<double> utmp;
+    };
+
+    template <class StokesMatrixType, class ABlockMatrixType, class SchurComplementMatrixType,
+              class ABlockPreconditionerType, class SchurComplementPreconditionerType>
+    BlockSchurGMGPreconditioner<StokesMatrixType, ABlockMatrixType, SchurComplementMatrixType,
+                                ABlockPreconditionerType, SchurComplementPreconditionerType>::
+                                BlockSchurGMGPreconditioner (const StokesMatrixType                  &Stokes_matrix,
+                                                             const ABlockMatrixType                  &A_block,
+                                                             const SchurComplementMatrixType         &Schur_complement_block,
+                                                             const ABlockPreconditionerType          &A_block_preconditioner,
+                                                             const SchurComplementPreconditionerType &Schur_complement_preconditioner,
+                                                             const bool                               do_solve_A,
+                                                             const bool                               do_solve_Schur_complement,
+                                                             const bool                               A_block_symmetric,
+                                                             const double                             A_block_tolerance,
+                                                             const double                             Schur_complement_tolerance)
+                                  :
+                                  stokes_matrix                   (Stokes_matrix),
+                                  A_block                         (A_block),
+                                  Schur_complement_block          (Schur_complement_block),
+                                  A_block_preconditioner          (A_block_preconditioner),
+                                  Schur_complement_preconditioner (Schur_complement_preconditioner),
+                                  do_solve_A                      (do_solve_A),
+                                  do_solve_Schur_complement       (do_solve_Schur_complement),
+                                  A_block_is_symmetric            (A_block_symmetric),
+                                  n_iterations_A_                 (0),
+                                  n_iterations_Schur_complement_  (0),
+                                  A_block_tolerance               (A_block_tolerance),
+                                  Schur_complement_tolerance      (Schur_complement_tolerance)
+    {}
+
+    template <class StokesMatrixType, class ABlockMatrixType, class SchurComplementMatrixType,
+              class ABlockPreconditionerType, class SchurComplementPreconditionerType>
+    unsigned int
+    BlockSchurGMGPreconditioner<StokesMatrixType, ABlockMatrixType, SchurComplementMatrixType,
+                                ABlockPreconditionerType, SchurComplementPreconditionerType>::
+                                n_iterations_A_block() const
+    {
+      return n_iterations_A_;
+    }
+
+    template <class StokesMatrixType, class ABlockMatrixType, class SchurComplementMatrixType,
+              class ABlockPreconditionerType, class SchurComplementPreconditionerType>
+    unsigned int
+    BlockSchurGMGPreconditioner<StokesMatrixType, ABlockMatrixType, SchurComplementMatrixType,
+                                ABlockPreconditionerType, SchurComplementPreconditionerType>::
+                                n_iterations_Schur_complement() const
+    {
+      return n_iterations_Schur_complement_;
+    }
+
+    template <class StokesMatrixType, class ABlockMatrixType, class SchurComplementMatrixType,
+              class ABlockPreconditionerType, class SchurComplementPreconditionerType>
+    void
+    BlockSchurGMGPreconditioner<StokesMatrixType, ABlockMatrixType, SchurComplementMatrixType,
+                                ABlockPreconditionerType, SchurComplementPreconditionerType>::
+                                vmult (dealii::LinearAlgebra::distributed::BlockVector<double>       &dst,
+                                       const dealii::LinearAlgebra::distributed::BlockVector<double>  &src) const
+    {
+      if (utmp.size()==0)
+        utmp.reinit(src);
+
+      // This needs to be done explicitly, as GMRES does not
+      // initialize the data of the vector dst before calling
+      // us. Otherwise we might use random data as our initial guess.
+      dst = 0.0;
+
+      // either solve with the Schur complement matrix (if do_solve_Schur_complement==true)
+      // or just apply one preconditioner sweep (for the first few
+      // iterations of our two-stage outer GMRES iteration)
+      if (do_solve_Schur_complement)
+        {
+          // first solve with the bottom right block, which we have built
+          // as a mass matrix with the inverse of the viscosity
+          SolverControl solver_control(100, src.block(1).l2_norm() * Schur_complement_tolerance,true);
+
+          SolverCG<dealii::LinearAlgebra::distributed::Vector<double>> solver(solver_control);
+          // Trilinos reports a breakdown
+          // in case src=dst=0, even
+          // though it should return
+          // convergence without
+          // iterating. We simply skip
+          // solving in this case.
+          if (src.block(1).l2_norm() > 1e-50)
+            {
+              try
+                {
+                  solver.solve(Schur_complement_block,
+                               dst.block(1), src.block(1),
+                               Schur_complement_preconditioner);
+                  n_iterations_Schur_complement_ += solver_control.last_step();
+                }
+              // if the solver fails, report the error from processor 0 with some additional
+              // information about its location, and throw a quiet exception on all other
+              // processors
+              catch (const std::exception &exc)
+                {
+                  Utilities::throw_linear_solver_failure_exception("iterative (bottom right) solver",
+                                                                   "BlockSchurGMGPreconditioner::vmult",
+                                                                   std::vector<SolverControl> {solver_control},
+                                                                   exc,
+                                                                   src.block(0).get_mpi_communicator());
+                }
+            }
+        }
+      else
+        {
+          Schur_complement_preconditioner.vmult(dst.block(1),src.block(1));
+          n_iterations_Schur_complement_ += 1;
+        }
+
+      dst.block(1) *= -1.0;
+
+      {
+        // the u-block of dst only contains zeros
+        stokes_matrix.vmult(utmp, dst); // B^T
+        utmp.block(0) *= -1.0;
+        utmp.block(0) += src.block(0);
+      }
+
+      // now either solve with the top left block (if do_solve_A==true)
+      // or just apply one preconditioner sweep (for the first few
+      // iterations of our two-stage outer GMRES iteration)
+      if (do_solve_A == true)
+        {
+          SolverControl solver_control(1000, utmp.block(0).l2_norm() * A_block_tolerance);
+          PrimitiveVectorMemory<dealii::LinearAlgebra::distributed::Vector<double>> mem;
+
+          try
+            {
+              if (A_block_is_symmetric)
+                {
+                  SolverCG<dealii::LinearAlgebra::distributed::Vector<double>> solver(solver_control,mem);
+                  solver.solve(A_block, dst.block(0), utmp.block(0),
+                               A_block_preconditioner);
+                }
+              else
+                {
+                  // Use BiCGStab for non-symmetric matrices.
+                  // BiCGStab can also solve indefinite systems if necessary.
+                  // Do not compute the exact residual, as this
+                  // is more expensive, and we only need an approximate solution.
+                  SolverBicgstab<dealii::LinearAlgebra::distributed::Vector<double>>
+                  solver(solver_control,
+                         mem,
+                         SolverBicgstab<dealii::LinearAlgebra::distributed::Vector<double>>::AdditionalData(/*exact_residual=*/ false));
+                  solver.solve(A_block, dst.block(0), utmp.block(0),
+                               A_block_preconditioner);
+                }
+
+              n_iterations_A_ += solver_control.last_step();
+            }
+          // if the solver fails, report the error from processor 0 with some additional
+          // information about its location, and throw a quiet exception on all other
+          // processors
+          catch (const std::exception &exc)
+            {
+              Utilities::throw_linear_solver_failure_exception("iterative (top left) solver",
+                                                               "BlockSchurGMGPreconditioner::vmult",
+                                                               std::vector<SolverControl> {solver_control},
+                                                               exc,
+                                                               src.block(0).get_mpi_communicator());
+            }
+        }
+      else
+        {
+          A_block_preconditioner.vmult (dst.block(0), utmp.block(0));
+          n_iterations_A_ += 1;
+        }
+    }
+  }
 }
 
 #endif
